+++ resources/views/painel/post_categories/index.blade.php
@@ -5,13 +5,15 @@
         <h1>Categorias de Posts</h1>
     </div>
 
-    <div class='col-md-12 text-center'>
-        <a href='{{route('post_categories.create')}}' alt='Cadastrar' title='Cadastrar' class='btn btn-default'>
-            Cadastrar
-        </a>
-        <br/>
-        <br/>
-    </div>
+    @can('create-post_categories')
+        <div class='col-md-12 text-center'>
+            <a href='{{route('post_categories.create')}}' alt='Cadastrar' title='Cadastrar' class='btn btn-default'>
+                Cadastrar
+            </a>
+            <br/>
+            <br/>
+        </div>
+    @endcan
 
     <table class="table table-responsive table-striped table-bordered table-hovered">
         <thead>
@@ -25,13 +27,17 @@
             @forelse ($postCategories as $postCategory)
             <tr>
                 <td>
-                    {!! Form::open(['route' => ['post_categories.destroy', $postCategory->id], 'method' => 'delete', 'style' => 'display: inline']) !!}
-                    {!! Form::button('<i class="fa fa-trash" aria-hidden="true"></i>', ['type' => 'submit', 'class' => 'btn btn-danger']) !!}
-                    {!! Form::close() !!}
-
-                    <a href='post_categories/{{$postCategory->id}}/edit' class='btn btn-info'>
-                        <i class="fa fa-pencil" aria-hidden="true"></i>
-                    </a>
+                    @can('delete-post_categories')
+                        {!! Form::open(['route' => ['post_categories.destroy', $postCategory->id], 'method' => 'delete', 'style' => 'display: inline']) !!}
+                            {!! Form::button('<i class="fa fa-trash" aria-hidden="true"></i>', ['type' => 'submit', 'class' => 'btn btn-danger']) !!}
+                        {!! Form::close() !!}
+                    @endcan
+                    
+                    @can('update-post_categories')
+                        <a href='post_categories/{{$postCategory->id}}/edit' class='btn btn-info'>
+                            <i class="fa fa-pencil" aria-hidden="true"></i>
+                        </a>
+                    @endcan
                 </td>
                 <td>{{$postCategory->id}}</td>
                 <td>{{$postCategory->title}}</td>
